Successfully vectorised the hot loop, BUT...
Writing the code was super buggy and it is confusing as hell to
understand.

Perhaps this is one of the rare situations where breaking it out into an
einsum makes the most sense.
I'm not sure where the time is still going: perhaps even in the exp(img
dot product) construction.