Upset with the size of Images.jl, I wrote my own. Hello PPM, old friend.
But seriously - PPM is such a simple format (with both ASCII and binary
forms), that it often makes a lot of sense to do this. All my C codes
form undergraduate used to hand-roll the output, and it also makes a lot
of sense in today's modern Julia world to avoid the heavyweight
Images.jl dependency.

Imagemagick convert is perfectly capable of converting from PPM to PNG.
(Some web browsers & etc. cannot display PPM directly.)